Split xrdp_bitmap.c into separate files for testing
